implement cache clearing for map tile cache
Cache Clearing runs once on app start. If on Initialization the number
of documents in the Map Tile Cache exceeds defined Threshold, the 1/3
oldest Tiles are deleted from the Cache.

The Size of the cache can be adjusted dynamically via the `const
maxDocCount` constant. The size of the docs itself is neglegible, each
doc has an image attached that is roughly 70KB in size. Currently 200
Elements are allowed (~11MB Image data).

Note: The caching does not work in Firefox if the option
`privacy.resistFingerprinting` is enabled, as it prevents the app from
extracting the tile data for storage. This is not afaik detectable by
the app and results in empty Tiles being displayed.